pm_automation_suite/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
dataclasses-json>=0.6.0

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# API integrations
atlassian-python-api>=3.41.0  # For Jira/Confluence
google-api-python-client>=2.100.0  # For Google Suite
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
snowflake-connector-python>=3.5.0  # For Snowflake
slack-sdk>=3.23.0  # For Slack integration (future)

# AI/ML libraries
openai>=1.0.0  # For OpenAI integration
anthropic>=0.25.0  # For Claude integration
langchain>=0.1.0  # For AI orchestration
tiktoken>=0.5.0  # For token counting
transformers>=4.35.0  # For local models (optional)
scikit-learn>=1.3.0  # For data analysis

# Data processing
openpyxl>=3.1.0  # For Excel support
pyarrow>=14.0.0  # For efficient data handling
fastparquet>=2023.10.0  # For parquet files

# Scheduling and orchestration
apscheduler>=3.10.0  # For task scheduling
celery>=5.3.0  # For distributed tasks (optional)
redis>=5.0.0  # For message broker (optional)

# Web framework (for API endpoints)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0

# Monitoring and logging
structlog>=23.0.0
rich>=13.0.0  # For pretty console output
prometheus-client>=0.19.0  # For metrics

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=20.0.0  # For test data generation

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0
pre-commit>=3.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocstrings>=0.24.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0
python-jose>=3.3.0
keyring>=24.0.0  # For secure credential storage

# Utilities
click>=8.1.0  # For CLI
tqdm>=4.66.0  # For progress bars
httpx>=0.25.0  # Modern HTTP client
tenacity>=8.2.0  # For retry logic
cachetools>=5.3.0  # For caching
python-dateutil>=2.8.0  # Date utilities
pytz>=2023.3  # Timezone support
validators>=0.22.0  # Data validation

# Optional performance improvements
# uvloop>=0.19.0  # Faster event loop (Unix only)
# orjson>=3.9.0  # Faster JSON parsing